Fix compiler warnings

+++ cpp/libcore/source/geometry/length_unit.hpp
@@ -189,53 +189,53 @@ inline auto operator*(double p_scalar, jps::LengthUnit p_lu) -> jps::LengthUnit
 /// User defined literals for all units
 inline auto operator"" _um(long double p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::um>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::um>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _mm(long double p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::mm>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::mm>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _cm(long double p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::cm>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::cm>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _dm(long double p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::dm>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::dm>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _m(long double p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::m>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::m>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _km(long double p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::km>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::km>(static_cast<double>(p_quantity));
 }
 
 /// User defined literals for integrals
 inline auto operator"" _um(unsigned long long p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::um>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::um>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _mm(unsigned long long p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::mm>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::mm>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _cm(unsigned long long p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::cm>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::cm>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _dm(unsigned long long p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::dm>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::dm>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _m(unsigned long long p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::m>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::m>(static_cast<double>(p_quantity));
 }
 inline auto operator"" _km(unsigned long long p_quantity) -> jps::LengthUnit
 {
-    return jps::makeLengthUnit<jps::Units::km>(p_quantity);
+    return jps::makeLengthUnit<jps::Units::km>(static_cast<double>(p_quantity));
 }
 
 namespace fmt
